+++ HISTORY.md
@@ -7,6 +7,7 @@
   - `absvector`, `address->`, `<-address`, `intern`, `n->string`, `pos`, `string->n`, and `tlstr` now throw errors catchable by `trap-error` when type violations occur with their arguments.
   - `tlstr` now raises an error when applied to an empty string
   - `set` and `value` now require their first arugment to be a symbol
+  - `hd` and `tl` now raise an error when applied to non-lists
 
 ## 0.3.1 - January 24, 2013
 - No code changes. Updates Gemspec to refer to Shen 8.0.